% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_classes.R
\name{class_vecvec}
\alias{class_vecvec}
\title{Construct a vector of vectors}
\usage{
class_vecvec(x = list(), i = seq_len(sum(lengths(x))))
}
\arguments{
\item{x}{An unnamed list of arbitrary vectors.}

\item{loc}{A named list of value locations, with \code{i} identifying the vector index and \code{x} identifying the value index. By default, the order of appearance in \code{x} will be used.}

\item{class}{Name of subclass.}
}
\value{
A vector of vectors of class \code{vecvec}.
}
\description{
new_vecvec() constructs a new vector of vectors from a list of vectors. It is meant to be performant, and does not check the inputs for correctness in any way. It is only safe to use after a call to df_list(), which collects and validates the columns used to construct the data frame.
}
\examples{
# Create a vecvec prototype
class_vecvec()

# Construct a vecvec from a list of vectors
class_vecvec(list(letters, rnorm(10)))

# Fully specify a vecvec with locations
class_vecvec(
  x = list(letters, rnorm(10)),
  i = c(1:3, 27:31, 26:4, 32:36)
)

}
