% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecvec.R
\name{create_match_matrix}
\alias{create_match_matrix}
\title{Create Match Matrix}
\usage{
create_match_matrix(.vecvec)
}
\arguments{
\item{.vecvec}{A \code{vecvec} object.}
}
\value{
A data frame containing the match matrix.
}
\description{
Creates a match matrix from a \code{vecvec} object.
}
