% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-names.R
\name{create_random_names}
\alias{create_random_names}
\title{Create Random Names}
\usage{
create_random_names(num_names = 100)
}
\arguments{
\item{num_names}{The number of names to generate.}
}
\value{
A character vector of random names.
}
\description{
Generates a list of random names.
}
\examples{
create_random_names(num_names = 10)
}
