% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{cross_join}
\alias{cross_join}
\title{Cross Join}
\usage{
cross_join(.data, lookup)
}
\arguments{
\item{.data}{A data frame.}

\item{lookup}{A lookup data frame.}
}
\value{
A cross-joined data frame.
}
\description{
Performs a cross join between two data frames.
}
