% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{exact_substring_match}
\alias{exact_substring_match}
\title{Exact Substring Match}
\usage{
exact_substring_match(.data)
}
\arguments{
\item{.data}{A data frame of tokens.}
}
\value{
A data frame of matched substrings.
}
\description{
Matches substrings between two data frames.
}
