% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{exact_token_match}
\alias{exact_token_match}
\title{Exact Token Match}
\usage{
exact_token_match(.data)
}
\arguments{
\item{.data}{A data frame of tokens.}
}
\value{
A data frame of matched tokens.
}
\description{
Matches tokens exactly between two data frames.
}
