% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{exact_whole_match}
\alias{exact_whole_match}
\title{Exact Whole Match}
\usage{
exact_whole_match(.data, lookup)
}
\arguments{
\item{.data}{A data frame of primary strings.}

\item{lookup}{A data frame of lookup strings.}
}
\value{
A data frame of matched rows.
}
\description{
Performs an exact match of whole strings between two data frames.
}
