% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{format_as_tidytext}
\alias{format_as_tidytext}
\title{Format as Tidytext}
\usage{
format_as_tidytext(.data)
}
\arguments{
\item{.data}{A data frame of strings.}
}
\value{
A tokenized data frame.
}
\description{
Formats a data frame for tokenized text analysis.
}
