% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{fuzzy_match}
\alias{fuzzy_match}
\title{Fuzzy Match}
\usage{
fuzzy_match(primary, lookup, dist = 1)
}
\arguments{
\item{primary}{A data frame of primary strings.}

\item{lookup}{A data frame of lookup strings.}

\item{dist}{The maximum distance for fuzzy matching.}
}
\value{
A data frame of fuzzy matches.
}
\description{
Performs a fuzzy match between two data frames.
}
