% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-names.R
\name{generate_name}
\alias{generate_name}
\title{Generate Name}
\usage{
generate_name(num_names)
}
\arguments{
\item{num_names}{The number of names to generate.}
}
\value{
A character vector of generated names.
}
\description{
Generates random names with transformations.
}
