% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-names.R
\name{generate_name_type}
\alias{generate_name_type}
\title{Generate Name Type}
\usage{
generate_name_type(num_names)
}
\arguments{
\item{num_names}{The number of names to generate.}
}
\value{
A tibble with name types.
}
\description{
Randomly assigns a name type (first, last, or both).
}
