% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-names.R
\name{generate_random_transform}
\alias{generate_random_transform}
\title{Generate Random Transform}
\usage{
generate_random_transform(num_names)
}
\arguments{
\item{num_names}{The number of names to generate.}
}
\value{
A tibble with name transformations.
}
\description{
Randomly assigns a case transformation to names.
}
