% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grab_cols}
\alias{grab_cols}
\title{Grab Columns}
\usage{
grab_cols(x)
}
\arguments{
\item{x}{A data frame.}
}
\value{
A data frame with selected columns.
}
\description{
Extracts specific columns from a data frame.
}
