% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index}
\alias{index}
\title{Index}
\usage{
index(vec)
}
\arguments{
\item{vec}{A character vector.}
}
\value{
A list containing data and string indices.
}
\description{
Creates an index for a character vector.
}
