% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecvec.R
\name{is_vecvec}
\alias{is_vecvec}
\title{Test if an object is a vecvec}
\usage{
is_vecvec(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{class_vecvec}, \code{FALSE} otherwise.
}
\description{
Test if an object is a vecvec
}
\examples{
vv <- vecvec(1:3, letters)
is_vecvec(vv)
is_vecvec(1:3)

}
\seealso{
\code{\link[=vecvec]{vecvec()}} to create a vecvec object.
}
