% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{map_ids}
\alias{map_ids}
\title{Map Indicies to Data}
\usage{
map_ids(.data, unique_strings)
}
\arguments{
\item{.data}{A data frame of indices.}

\item{unique_strings}{A data frame of unique strings.}
}
\value{
A data frame with mapped IDs.
}
\description{
Maps unique string IDs back to the original data.
}
