% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic.R
\name{masked-cor}
\alias{masked-cor}
\alias{var}
\alias{cov}
\alias{cor}
\title{Correlation, Variance and Covariance}
\usage{
var(x, ...)

cov(x, ...)

cor(x, ...)
}
\arguments{
\item{x}{a numeric vector, matrix or data frame.}

\item{...}{Additional arguments passed to \code{\link[stats:var]{stats::var}}, \code{\link[stats:cov]{stats::cov}}, and
\code{\link[stats:cor]{stats::cor}}.}
}
\description{
\code{var}, \code{cov}, and \code{cor}  compute the variance of x and the covariance or
correlation of x and y if these are vectors. If x and y are matrices then
the covariances (or correlations) between the columns of x and the columns
of y are computed.

These functions mask and wrap the \code{\link[stats:var]{stats::var}}, \code{\link[stats:cov]{stats::cov}}, and
\code{\link[stats:cor]{stats::cor}} functions to add support for \code{vecvec} objects. More details
can be found in the documentation for those functions.
}
\seealso{
\code{\link[stats:var]{stats::var()}}, \code{\link[stats:cov]{stats::cov()}}, \code{\link[stats:cor]{stats::cor()}}
}
