% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match}
\alias{match}
\title{Match}
\usage{
match(primary, lookup)
}
\arguments{
\item{primary}{A data frame of primary strings.}

\item{lookup}{A data frame of lookup strings.}
}
\value{
A list of matches.
}
\description{
Matches primary strings against lookup strings using various methods.
}
