% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{prep_lookup}
\alias{prep_lookup}
\title{Prepare Lookup Tibble}
\usage{
prep_lookup(lookup)
}
\arguments{
\item{lookup}{A lookup tibble.}
}
\value{
A prepped lookup tibble.
}
\description{
Standardizes the lookup tibble.
}
