% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{prep_stnd_string}
\alias{prep_stnd_string}
\title{Prepare Standardized String}
\usage{
prep_stnd_string(.data)
}
\arguments{
\item{.data}{A data frame containing a \code{stnd_string} column.}
}
\value{
A data frame with prepared strings.
}
\description{
Prepares a standardized string for matching.
}
