% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{prep_string}
\alias{prep_string}
\title{Prepare String}
\usage{
prep_string(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A prepared character vector.
}
\description{
Prepares a string by replacing separators, removing slashes, and removing numbers.
}
