% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{reduce_stnd_strings}
\alias{reduce_stnd_strings}
\title{Reduce Standardized Strings}
\usage{
reduce_stnd_strings(.data)
}
\arguments{
\item{.data}{A data frame of standardized strings.}
}
\value{
A data frame of unique strings.
}
\description{
Reduces a data frame to unique standardized strings.
}
