% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{remove_all_spaces}
\alias{remove_all_spaces}
\title{Remove All Spaces}
\usage{
remove_all_spaces(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector without spaces.
}
\description{
Removes all spaces from a character vector.
}
