% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{remove_numbers}
\alias{remove_numbers}
\title{Remove Numbers}
\usage{
remove_numbers(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector without numbers.
}
\description{
Removes numbers from a character vector.
}
