% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{remove_slash}
\alias{remove_slash}
\title{Remove Slash}
\usage{
remove_slash(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector without slashes.
}
\description{
Removes slashes from a character vector.
}
