% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{standardize_str}
\alias{standardize_str}
\title{Standardize String}
\usage{
standardize_str(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A standardized character vector.
}
\description{
Converts a string to lowercase and removes extra spaces.
}
