% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{standardize_vec}
\alias{standardize_vec}
\title{Standardize Vector}
\usage{
standardize_vec(vec)
}
\arguments{
\item{vec}{A character vector.}
}
\value{
A tibble with standardized strings.
}
\description{
Standardizes a character vector for indexing.
}
