% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{tidy_matches}
\alias{tidy_matches}
\title{Tidy Matches}
\usage{
tidy_matches(primary, lookup)
}
\arguments{
\item{primary}{A data frame of primary strings.}

\item{lookup}{A data frame of lookup strings.}
}
\value{
A list of tidy matches.
}
\description{
Performs tidy matches (exact tokens, substrings, and fuzzies).
}
