% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecvec.R
\name{unique_matches}
\alias{unique_matches}
\title{Extract Unique Matches}
\usage{
unique_matches(
  match_matrix,
  match_vars = c("whole", "token", "substring", "fuzzy")
)
}
\arguments{
\item{match_matrix}{A match matrix.}

\item{match_vars}{A character vector of match variable names.}
}
\value{
A data frame of unique matches.
}
\description{
Extracts unique matches from a match matrix.
}
