% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecvec.R
\name{unvecvec}
\alias{unvecvec}
\title{Obtain a singular common vector type from a vector of vectors}
\usage{
unvecvec(x, ..., ptype = NULL)
}
\arguments{
\item{x}{A vecvec to unvecvec (convert to its underlying vector type)}
}
\value{
A simple vector, all containing the same type of data.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
