% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecvec-package.R
\docType{package}
\name{vecvec-package}
\alias{vecvec-package}
\title{vecvec: Construct Mixed Type Data Structures with Vectors of Vectors}
\description{
Mixed type vectors are useful for combining semantically similar classes. Some examples of semantically related classes include time across different granularities (e.g. daily, monthly, annual) and probability distributions (e.g. Normal, Uniform, Poisson). These groups of vector types typically share common statistical operations which vary in results with the attributes of each vector. The 'vecvec' data structure facilitates efficient storage and computation across multiple vectors within the same object.
}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com} (\href{https://orcid.org/0000-0001-6729-7695}{ORCID})

}
\keyword{internal}
