% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{vecvec_apply}
\alias{vecvec_apply}
\title{Apply a function to each vector of the vecvec}
\usage{
vecvec_apply(x, .f, ...)
}
\arguments{
\item{x}{A vecvec object}

\item{.f}{A function to apply to each vector}

\item{...}{Additional arguments passed to \code{.f}}
}
\value{
A vecvec data type with the same structure as \code{x} but with each
vector transformed by \code{.f}.
}
\description{
The \code{vecvec_apply()} function applies a function \code{.f} to each vector in the
\code{vecvec} vectors.
}
