% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hook.R
\name{vecvec_register}
\alias{vecvec_register}
\title{Register methods for a vecvec subclass}
\usage{
vecvec_register(
  x,
  f_ptype2 = vec_ptype2_vecvec,
  f_cast_to = vec_cast_to_vecvec,
  f_cast_from = vec_cast_from_vecvec
)
}
\arguments{
\item{x}{An S7 class object that extends \link{class_vecvec}. Typically the
class object created by \code{\link[S7:new_class]{S7::new_class()}} in your package, e.g.
\code{class_myvec}.}

\item{f_ptype2}{The function to register as the \code{vec_ptype2} method for this
class. Defaults to \code{vec_ptype2_vecvec}, guarantees that the common type of
two vectors is a vecvec vector.}

\item{f_cast_to}{The function to register as the \verb{vec_cast.<class>.*} methods
for this class. Defaults to \code{vec_cast_to_vecvec}, which attempts to cast
other types to vecvec by calling \code{vecvec()}.}

\item{f_cast_from}{The function to register as the \verb{vec_cast.*.<class>}
methods for this class. Defaults to \code{vec_cast_from_vecvec}, which attempts
to cast vecvec to other types by calling \code{unvecvec()}.}
}
\value{
\code{NULL}, invisibly. Called for its side effects.
}
\description{
Call \code{vecvec_register()} inside your package's \code{.onLoad()} function to
register the S3 methods that allow a custom
\link[=class_vecvec]{vecvec} subclass to participate in the
\href{https://vctrs.r-lib.org/}{vctrs} vector system.
}
\details{
Specifically, \code{vecvec_register()} registers:
\itemize{
\item \verb{vec_cast.<class>.*} methods, so that other types can be cast \emph{to} your
vecvec subclass.
\item \verb{vec_cast.*.<class>} methods, so that your vecvec subclass can be cast
\emph{from} other types.
}
}
\examples{
\dontrun{
# In your package, define a vecvec subclass:
class_myvec <- S7::new_class(
  "myvec",
  parent = vecvec::class_vecvec
)

# Then register it in .onLoad():
.onLoad <- function(libname, pkgname) {
  S7::methods_register()
  vecvec::vecvec_register(class_myvec)
}
}
}
