% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecvec.R
\name{human_readable}
\alias{human_readable}
\title{Human-Readable Matches}
\usage{
human_readable(.vecvec, match_vars = c("whole", "token", "substring", "fuzzy"))
}
\arguments{
\item{.vecvec}{A \code{vecvec} object.}

\item{match_vars}{A character vector of match variable names.}
}
\value{
A data frame of human-readable matches.
}
\description{
Converts a \code{vecvec} object into a human-readable format.
}
