% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecvec.R
\name{new_vecvec}
\alias{new_vecvec}
\title{Create a new vector of vectors}
\usage{
new_vecvec(...)
}
\arguments{
\item{...}{Individual vectors to combine into a vecvec}
}
\description{
Create a new vector of vectors
}
\examples{
new_vecvec(Sys.Date(), rnorm(3), letters)

}
