% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecvec-package.R
\docType{package}
\name{vecvec-package}
\alias{vecvec}
\alias{vecvec-package}
\title{vecvec: Construct Mixed Type Data Structures with Vectors of Vectors}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com} (\href{https://orcid.org/0000-0001-6729-7695}{ORCID})

}
\keyword{internal}
