% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecvec.R
\name{vecvec}
\alias{vecvec}
\title{Match Primary Vector Against Lookup Vector}
\usage{
vecvec(primary, lookup)
}
\arguments{
\item{primary}{A character vector to be matched.}

\item{lookup}{A character vector to match against.}
}
\value{
A list containing indices and matches.
}
\description{
Matches a primary character vector against a lookup vector and returns indices and matches.
}
\examples{
vecvec(create_random_names(), create_random_names())
}
