% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecvec.R
\name{vecvec}
\alias{vecvec}
\title{Create a new vector of vectors}
\usage{
vecvec(..., class = character())
}
\arguments{
\item{...}{Vectors to combine into a single vector without type coercion.}

\item{class}{Name of subclass.}
}
\value{
A vector of vectors of class \code{vecvec}.
}
\description{
Create a new vector of vectors
}
\examples{
vecvec(Sys.Date(), rnorm(3), letters)

}
\seealso{
\code{\link[=unvecvec]{unvecvec()}} coerces the mixed-type vector into a single-typed
regular vector. \code{\link[=new_vecvec]{new_vecvec()}} is a performant alternative that accepts a
list of vectors rather than \code{...} (suitable for R packages).
}
