% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecvec_apply.R
\name{vecvec_apply}
\alias{vecvec_apply}
\title{Apply a function to each vector of the vecvec}
\usage{
vecvec_apply(x, f, ...)
}
\arguments{
\item{x}{A vecvec object}

\item{f}{A function to apply to each vector}

\item{...}{Additional arguments passed to \code{f}}
}
\description{
Apply a function to each vector of the vecvec
}
